Update Starfish to use Float data Exclusively (#501)
* ENH: convert pipeline to working with float32 in [0, 1]
* ENH: make fix to richardson-lucy
* Update MERFISH notebook with new threshold

+++ starfish/test/intensity_table/test_intensity_table_serialization.py
@@ -17,7 +17,7 @@ def test_intensity_table_serialization():
     """
 
     # create an IntensityTable
-    data = np.zeros(100).reshape(1, 5, 2, 2, 5)
+    data = np.zeros(100, dtype=np.float32).reshape(1, 5, 2, 2, 5)
     image_stack = ImageStack.from_numpy_array(data)
     intensities = IntensityTable.from_image_stack(image_stack)
